Don't need to create thread as "daemon"
It raises a concern for @ThomasWaldmann because
"I first used also daemon mode. Somehow I noticed later that
there were a lot of threads (more than expected) and the fact
that it didn't work as expected was hidden by using daemon
mode.

"I switched then to non-daemon mode (which means you have to
cleanly tear down the threads you create)."

My code is of course perfect and cleans up every time.  But
we have a lovely testsuite which will regression-test this
property for us :p.

It also saves readers learning a new API feature, and makes
the code more explicit, so I agree with this suggestion.